[OPP-1331] lage multi-module prosjekt med modialogin og hosting-app
Repoet bygger to artifakter;
- modialogin, en app som deployes til preprod og prod. Appen har ansvar for å gjøre innlogging av en bruker. Erstatter bruken av veilarblogin
- modialogin-frontend, docker-image som kan brukes for en frontend-app som trenger autentisering. Erstatter bruken av pus-fss-frontend

Kodebasene vi erstatter har tidligere vært skilt fra hverandre, men siden dette er to "komponenter" som må samarbeide tett er de lagt inn her i en mono-repo struktur, med oppsett for å enkelt kunne bygge ut tester av hele verdikjeden.

En enkel måte å starte opp hele løsningen på er;
```
./gradlew clean build
docker-compose up --build
```

Man vil da kunne gå til http://localhost:8082/modiafrontend Og om man sjekker nettverksloggen (eller cookies) vil man se at man har blitt logget inn via modialogin og oidc-stub.

Om man jobber med videreutvikling av løsningene har man også mulighet til å starte applikasjonene enkeltvist via henholdsvis:
- `oidc-stub/OidcStub.kt`
- `login-app/MainTest.kt`
- `frontend-app/MainTest.kt`
