+++ .github/workflows/frontend-app-workflow.yml
@@ -1,0 +1,64 @@
+name: Frontend app
+
+on:
+  push:
+    paths-ignore:
+      - 'oidc-stub/**'
+      - 'login-app/**'
+      - '**.md'
+      - '.editorconfig'
+      - '.gitignore'
+      - 'CODEOWNERS'
+      - 'docker-compose.yml'
+
+env:
+  IMAGE: docker.pkg.github.com/${{ github.repository }}/modialogin-frontend:${{ github.sha }}
+  CI: true
+  TZ: Europe/Oslo
+jobs:
+  build:
+    name: Test and build
+    runs-on: ubuntu-latest
+    if: github.ref != 'refs/heads/dev' && github.ref != 'refs/heads/master'
+    steps:
+      - uses: actions/checkout@v1
+      - name: Set up JDK 11
+        uses: actions/setup-java@v1
+        with:
+          java-version: 11
+      - uses: actions/cache@v1
+        with:
+          path: ~/.gradle
+          key: ${{ runner.os }}-gradle-${{ hashFiles('**/build.gradle.kts') }}
+          restore-keys: |
+            ${{ runner.os }}-gradle-
+          continue-on-error: true
+      - name: Build
+        run:  ./gradlew build
+
+  build-and-push:
+    name: Build and push Docker container
+    if: github.ref == 'refs/heads/dev' || github.ref == 'refs/heads/master'
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v1
+      - name: Set up JDK 11
+        uses: actions/setup-java@v1
+        with:
+          java-version: 11
+      - uses: actions/cache@v1
+        with:
+          path: ~/.gradle
+          key: ${{ runner.os }}-gradle-${{ hashFiles('**/build.gradle.kts') }}
+          restore-keys: |
+            ${{ runner.os }}-gradle-
+        continue-on-error: true
+      - name: Build
+        run: ./gradlew build
+      - name: Publish Docker image
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        run: |
+          docker login docker.pkg.github.com -u ${GITHUB_REPOSITORY} -p ${GITHUB_TOKEN}
+          docker build --tag ${IMAGE} login-app/
+          docker push ${IMAGE}
