--- .github/workflows/main.yml
@@ -1,86 +1,0 @@
-name: Build, push, and deploy
-
-on: [push]
-
-env:
-  IMAGE: docker.pkg.github.com/${{ github.repository }}/modialogin:${{ github.sha }}
-  CI: true
-  TZ: Europe/Oslo
-jobs:
-  build:
-    name: Test and build
-    runs-on: ubuntu-latest
-    if: github.ref != 'refs/heads/dev' && github.ref != 'refs/heads/master'
-    steps:
-      - uses: actions/checkout@v1
-      - name: Set up JDK 11
-        uses: actions/setup-java@v1
-        with:
-          java-version: 11
-      - uses: actions/cache@v1
-        with:
-          path: ~/.gradle
-          key: ${{ runner.os }}-gradle-${{ hashFiles('**/build.gradle.kts') }}
-          restore-keys: |
-            ${{ runner.os }}-gradle-
-          continue-on-error: true
-      - name: Build
-        run:  ./gradlew build
-
-  build-and-push:
-    name: Build and push Docker container
-    if: github.ref == 'refs/heads/dev' || github.ref == 'refs/heads/master'
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v1
-      - name: Set up JDK 11
-        uses: actions/setup-java@v1
-        with:
-          java-version: 11
-      - uses: actions/cache@v1
-        with:
-          path: ~/.gradle
-          key: ${{ runner.os }}-gradle-${{ hashFiles('**/build.gradle.kts') }}
-          restore-keys: |
-            ${{ runner.os }}-gradle-
-        continue-on-error: true
-      - name: Build
-        run: ./gradlew build
-      - name: Publish Docker image
-        env:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-        run: |
-          docker login docker.pkg.github.com -u ${GITHUB_REPOSITORY} -p ${GITHUB_TOKEN}
-          docker build --tag ${IMAGE} .
-          docker push ${IMAGE}
-  deploy-qa:
-    name: Deploy to preprod
-    needs: build-and-push
-    if: github.ref == 'refs/heads/dev'
-    runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        namespace: [ q0, q1 ]
-    steps:
-      - uses: actions/checkout@v1
-      - uses: nais/deploy/actions/deploy@v1
-        env:
-          APIKEY: ${{ secrets.NAIS_DEPLOY_APIKEY }}
-          PRINT_PAYLOAD: true
-          CLUSTER: dev-fss
-          RESOURCE: .nais/preprod.yml
-          VAR: namespace=${{ matrix.namespace }}
-
-  deploy-prod:
-    name: Deploy to prod
-    needs: build-and-push
-    if: github.ref == 'refs/heads/master'
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v1
-      - uses: nais/deploy/actions/deploy@v1
-        env:
-          APIKEY: ${{ secrets.NAIS_DEPLOY_APIKEY }}
-          PRINT_PAYLOAD: true
-          CLUSTER: prod-fss
-          RESOURCE: .nais/prod.yml
