+++ common/src/main/kotlin/no/nav/modialogin/common/ModiaLoginConstants.kt
@@ -1,0 +1,4 @@
+object ModiaLoginConstants {
+    const val idToken = "modia_ID_token"
+    const val refreshToken = "modia_refresh_token"
+}
