+++ common/src/main/kotlin/no/nav/modialogin/common/features/NaisFeature.kt
@@ -1,0 +1,30 @@
+package no.nav.modialogin.common.features
+
+import io.ktor.application.*
+import io.ktor.http.*
+import io.ktor.response.*
+import io.ktor.routing.*
+import no.nav.modialogin.common.NaisState
+
+fun Application.installNaisFeature(appname: String, config: NaisState) {
+    routing {
+        route(appname) {
+            route("internal") {
+                get("isAlive") {
+                    if (config.isAlive) {
+                        call.respondText("Alive: $appname")
+                    } else {
+                        call.respondText("Not alive", status = HttpStatusCode.InternalServerError)
+                    }
+                }
+                get("isReady") {
+                    if (config.isReady) {
+                        call.respondText("Ready")
+                    } else {
+                        call.respondText("Not ready", status = HttpStatusCode.InternalServerError)
+                    }
+                }
+            }
+        }
+    }
+}
