+++ frontend-app/index.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Testpage for modiafrontend</title>
+</head>
+<body>
+<h1>Testpage for modiafrontend</h1>
+</body>
+</html>
