+++ login-app/Dockerfile
@@ -1,0 +1,3 @@
+FROM docker.pkg.github.com/navikt/pus-nais-java-app/pus-nais-java-app:java11
+
+COPY login-app/build/libs/app.jar app.jar
