+++ login-app/src/main/kotlin/no/nav/modialogin/LoginAppConfig.kt
@@ -1,0 +1,14 @@
+package no.nav.modialogin
+
+import dev.nohus.autokonfig.types.BooleanSetting
+import dev.nohus.autokonfig.types.IntSetting
+import dev.nohus.autokonfig.types.StringSetting
+
+class LoginAppConfig {
+    val appname by StringSetting()
+    val idpDiscoveryUrl by StringSetting()
+    val idpClientId by StringSetting()
+    val idpClientSecret by StringSetting()
+    val xForwardingPort by IntSetting(default = 8080)
+    val dockerCompose by BooleanSetting(default = false)
+}
