+++ oidc-stub/build.gradle.kts
@@ -1,0 +1,61 @@
+import org.gradle.jvm.tasks.Jar
+import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
+
+val ktorVersion = "1.5.2"
+val logbackVersion = "1.2.3"
+val logstashVersion = "5.1"
+
+plugins {
+    application
+    kotlin("jvm") version "1.4.30"
+    kotlin("plugin.serialization") version "1.4.30"
+}
+
+group = "no.nav"
+version = ""
+
+repositories {
+    mavenCentral()
+    jcenter()
+}
+
+dependencies {
+    implementation(kotlin("stdlib"))
+    implementation("io.ktor:ktor-server-netty:$ktorVersion")
+    implementation("io.ktor:ktor-serialization:$ktorVersion")
+    implementation("com.nimbusds:nimbus-jose-jwt:9.7")
+    implementation("ch.qos.logback:logback-classic:$logbackVersion")
+}
+
+java {
+    sourceCompatibility = JavaVersion.VERSION_11
+    targetCompatibility = JavaVersion.VERSION_11
+}
+tasks.withType<KotlinCompile> {
+    kotlinOptions.jvmTarget = "11"
+}
+
+tasks.test {
+    useJUnitPlatform()
+    testLogging {
+        events("passed", "skipped", "failed")
+    }
+}
+
+val fatJar = task("fatJar", type = Jar::class) {
+    archiveBaseName.set("app")
+    duplicatesStrategy = DuplicatesStrategy.INCLUDE
+    manifest {
+        attributes["Implementation-Title"] = "ModiaLoginStub"
+        attributes["Implementation-Version"] = archiveVersion
+        attributes["Main-Class"] = "no.nav.modialogin.OidcStubKt"
+    }
+    from(configurations.runtimeClasspath.get().map { if (it.isDirectory) it else zipTree(it) })
+    with(tasks.jar.get() as CopySpec)
+}
+
+tasks {
+    "build" {
+        dependsOn(fatJar)
+    }
+}
