+++ settings.gradle.kts
@@ -1,1 +1,5 @@
 rootProject.name = "modialogin"
+include("common")
+include("login-app")
+include("frontend-app")
+include("oidc-stub")
