--- src/main/kotlin/no/nav/modialogin/Main.kt
@@ -1,5 +1,0 @@
-package no.nav.modialogin
-
-fun main() {
-    startApplication()
-}
